Improved file structure, added overridable empty methods to classes
