+++ js/controllers/back-to-start.js
@@ -1,7 +1,7 @@
-import {initialState, views, recordedAnswers} from '../../data/gamedata';
-import BackToStartView from './back-to-start-view';
-import insertTemplate from '../../utils/insert-template';
-import renderGreeting from '../greeting';
+import {initialState, views, recordedAnswers} from '../data/gamedata';
+import BackToStartView from '../views/partials/back-to-start-view';
+import insertTemplate from '../utils/insert-template';
+import renderGreeting from './greeting';
 
 
 export default () => {
