+++ js/controllers/info-bar.js
@@ -1,0 +1,7 @@
+import InfoBarView from '../views/partials/info-bar-view';
+
+export default (state) => {
+  const infoBar = new InfoBarView(state);
+
+  return infoBar.element;
+};
