+++ js/controllers/stats.js
@@ -1,0 +1,8 @@
+import StatsView from '../views/stats-view';
+
+
+export default (state) => {
+  const stats = new StatsView(state);
+
+  return stats.element;
+};
