+++ js/main.js
@@ -1,5 +1,5 @@
 import {initialState} from './data/gamedata';
 import insertTemplate from './utils/insert-template';
-import renderIntro from './templates/intro';
+import renderIntro from './controllers/intro';
 
 insertTemplate(renderIntro(initialState));
