+++ js/views/game-view.js
@@ -1,7 +1,7 @@
 import {TYPE_RADIO_1, TYPE_RADIO_2, TYPE_PICTURE} from '../data/gamedata';
-import AbstractView from '../utils/abstract-view';
-import renderInfoBar from './partials/info-bar';
-import renderStatusBar from './partials/status-bar';
+import AbstractView from './abstract-view';
+import renderInfoBar from '../controllers/info-bar';
+import renderStatusBar from '../controllers/status-bar';
 import getRandomPic from '../utils/pic-randomizer';
 
 
@@ -34,6 +34,8 @@ export default class GameView extends AbstractView {
       formElem.reset();
     };
   }
+
+  onFormClick(evt) {}
 }
 
 /**
