+++ js/views/greeting-view.js
@@ -1,4 +1,4 @@
-import AbstractView from '../utils/abstract-view';
+import AbstractView from './abstract-view';
 
 export default class GreetingView extends AbstractView {
   get template() {
@@ -29,4 +29,6 @@ export default class GreetingView extends AbstractView {
       this.onStatsBtnClick();
     };
   }
+
+  onProceedBtnClick() {}
 }
