+++ js/views/stats-view.js
@@ -1,11 +1,11 @@
 import {rates} from '../data/gamedata';
-import AbstractView from '../utils/abstract-view';
-import renderBackBtn from './partials/back-to-start';
-import renderStatusBar from './partials/status-bar';
+import AbstractView from '../views/abstract-view';
+import renderBackBtn from '../controllers/back-to-start';
+import renderStatusBar from '../controllers/status-bar';
 import getResults from '../utils/calculate-score';
 
 
-class StatsView extends AbstractView {
+export default class StatsView extends AbstractView {
   constructor(state) {
     super();
     this._state = state;
@@ -60,9 +60,3 @@ class StatsView extends AbstractView {
     statsContainer.appendChild(renderStatusBar(this._state));
   }
 }
-
-export default (state) => {
-  const stats = new StatsView(state);
-
-  return stats.element;
-};
